feat(mapper): add method to map list of Paciente to DTOs
This commit introduces a new method `toDto` in the `PacienteMapper` interface to handle mapping of a list of `Paciente` entities to a list of `PacienteDto` objects. This enhancement improves the mapper's functionality by supporting bulk conversions, which is useful for scenarios where multiple patient records need to be processed at once.